+++ etna/analysis/outliers/density_outliers.py
@@ -2,6 +2,7 @@ from typing import TYPE_CHECKING
 from typing import Callable
 from typing import Dict
 from typing import List
+from typing import Union
 
 import numpy as np
 import pandas as pd
@@ -91,7 +92,7 @@ def get_anomalies_density(
     distance_coef: float = 3,
     n_neighbors: int = 3,
     distance_func: Callable[[float, float], float] = absolute_difference_distance,
-) -> Dict[str, List[pd.Timestamp]]:
+) -> Dict[str, List[Union[pd.Timestamp, int]]]:
     """Compute outliers according to density rule.
 
     For each element in the series build all the windows of size ``window_size`` containing this point.